+++ backend/src/main/java/com/shinhan/shbhack/ijoa/api/controller/diary/DiaryController.java
@@ -44,10 +44,11 @@ public class DiaryController {
 
 
 //    @PostMapping(value="", consumes = { MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE })
-@PostMapping("/")
+@PostMapping("")
  public ResponseEntity<?> writeDiary(@RequestPart(value = "info") DiaryCreateRequest diaryCreateRequest,
                                         @RequestPart(value="images", required = false)List<MultipartFile> images,
                                         @RequestPart(value="record",required = false)MultipartFile record){
+
         diaryCreateRequest.setPhoto(images);
         diaryCreateRequest.setRecord(record);
         diaryService.writeDiary(new DiaryCreateServiceRequest(diaryCreateRequest));
