+++ backend/src/main/java/com/shinhan/shbhack/ijoa/domain/bank/repository/datajpa/AccountRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AccountRepository  extends JpaRepository<Account, Long> {
     public Account findAccountByAccountNumber(String accountNumber);
-
+    public boolean existsByAccountNumber(String accountNumber);
 }
